Moved settings to separate file
Settings are now in settings.txt, which is created at the first
execution. Lines in the file are loaded individually, and each one is
cut off by the first '#' symbol and trimmed. If anything is left, it is
considered a valid line and loaded. Settings must be in order. Also
fixed issues with stopping the program. It should now correctly stop
when told to do so.